app/requirements.txt
asgiref==3.8.1
boto3==1.35.54
botocore==1.35.54
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
crispy-bootstrap5==0.7
cryptography==43.0.3
defusedxml==0.8.0rc2
django==5.1.3
django-filter==23.3 
django-allauth==0.54.0
django-cors-headers==4.6.0
django-crispy-forms==2.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-debug-toolbar==5.0.1
djoser==2.3.1
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
oauthlib==3.2.2
packaging==24.2
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.3
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
urllib3==2.2.3
redis==5.0.0
hiredis==2.2.3
django-redis==5.3.0
drf-yasg==1.21.8
drf-spectacular==0.26.4
django-import-export
celery==5.4.0
django-celery-beat==2.7.0